Refactor offer creation logic into service layer with improved error handling
