+++ src/controllers/offer.controller.ts
@@ -156,463 +156,27 @@ export async function getCounterOffer(request: FastifyRequest, reply: FastifyRep
 }
 
 export async function createOffer(request: FastifyRequest, reply: FastifyReply) {
-  const { program_id } = request.params as { program_id: string };
-  const offer = request.body as OfferInterface;
-  const traceId = generateCustomUUID();
-  const authHeader = request.headers.authorization;
-  const transaction: Transaction = await sequelize.transaction();
-  const user = request.user
-  const user_id = user?.sub
-  const userData = await offerRepository.findUser(program_id, user_id);
-  if (user?.userType !== "super_user" && (!userData || userData.length === 0)) {
-    await transaction?.rollback();
-    return reply.status(400).send({
-      status_code: 400,
-      trace_id: traceId,
-      message: "User not found",
-    });
-  }
-  let userTypes = user?.userType ?? userData[0]?.user_type ?? "";
   try {
-    let jobs = await JobModel.findOne({ where: { id: offer.job_id }, transaction });
-
-    if (!jobs) {
-      return reply.status(400).send({
-        status_code: 400,
-        message: "Job not found",
-        trace_id: traceId,
-      });
-    }
-    const holdStatuses = ["HOLD", "PENDING_REVIEW", "DRAFT", "FILLED", "CLOSED", "PENDING_APPROVAL", "REJECTED"];
-    const jobId = jobs.id
-    if (userTypes === "vendor") {
-      await checkDistributionStatusForVendor(program_id, userTypes, jobId, userData, reply, traceId)
-    }
-    if (holdStatuses.includes(jobs?.status)) {
-      reply.status(400).send({
-        status_code: 400,
-        message: `Job is currently on ${jobs?.status} offer cannot be created.`,
-        trace_id: traceId,
-      });
-      return;
-    }
-
-    if (!authHeader?.startsWith('Bearer ')) {
-      return reply.status(401).send({ message: 'Unauthorized - Token not found' });
-    }
-
-    const token = authHeader.split(' ')[1];
-
-    const user = request.user;
-    const userId = user.sub;
-    const userType = user.userType;
-
-    const offerData: any = request.body as OfferInterface;
-
-    const offerStartDate = new Date(offer.start_date);
-    const offerEndDate = new Date(offer.end_date);
-
-    if (offerStartDate > offerEndDate) {
-      return reply.status(400).send({
-        message: `Offer end date must be greater than the offer start date.`,
-      });
-    }
-
-    if (!offerData.financial_details?.rates) {
-      return reply.status(400).send({ message: "Financial details cannot be null" });
-    }
-
-    if (!offerData.financial_details?.fee_details) {
-      return reply.status(400).send({ message: "Fee details cannot be null" });
-    }
-
-    const submission = await SubmissionCandidateModel.findOne(
-      {
-        where: {
-          candidate_id: offerData.candidate_id,
-          program_id: program_id,
-          job_id: offerData.job_id
-        },
-        transaction
-      })
-
-    const validStatuses = [
-      Status.PENDING_REHIRE_APPROVAL,
-      Status.PENDING_REHIRE_REVIEW,
-      Status.PENDING_SHORTLIST_REVIEW,
-      Status.WITHDRAW,
-      Status.WITHDRAWN
-    ];
-    const submissionStatus = submission?.dataValues?.status?.toUpperCase() ?? "";
-
-    if (validStatuses.includes(submissionStatus)) {
-      return reply.status(400).send({
-        status_code: 400,
-        message: `Cannot create offer. Submission is currently in '${submissionStatus}' status.`,
-        trace_id: traceId,
-      });
-    }
-
-
-    const query = `SELECT * FROM jobs WHERE id = :jobId LIMIT 1;`;
-    const jobRequest: any = await sequelize.query(query, {
-      type: QueryTypes.SELECT,
-      replacements: { jobId: offerData.job_id },
+    const result = await offerService.createOffer(request, reply);
+    
+    return reply.status(result.status_code).send({
+      status_code: result.status_code,
+      trace_id: result.trace_id,
+      message: result.message,
+      ...(result.id && { id: result.id }),
+      ...(result.error && { error: result.error })
     });
-
-    let jobDatas = jobRequest[0];
-    if (!jobDatas) {
-      return reply.status(400).send({ massage: "Job not found for the provided job ID." });
-    }
-
-    const workflow_job_id = offerData.job_id;
-    const event_slug_base = "create_offer";
-    const module_name_base = "Offers";
-    const type = "workflow";
-    const placement_order = "0";
-    const is_updated = false;
-
-    const parent_offer_id = offerData?.parent_offer_id;
-    const event_slug = parent_offer_id ? "counter_offer" : event_slug_base;
-    const module_name = module_name_base;
-    let moduleId: any;
-
-    if (module_name) {
-      const query = `SELECT id FROM ${config_db}.module WHERE name = :module_name AND is_workflow = true LIMIT 1`;
-      let moduleIds = await sequelize.query(query, {
-        type: QueryTypes.SELECT,
-        replacements: { module_name },
-      });
-      moduleId = moduleIds[0];
-    }
-
-    const module_ids = moduleId?.id ?? "";
-    let eventId: any;
-    if (module_ids && event_slug) {
-      const query = `SELECT id FROM ${config_db}.event WHERE module_id = :module_ids AND slug = :event_slug AND is_enabled = true AND type = :type LIMIT 1`;
-      const eventIdData = await sequelize.query(query, {
-        type: QueryTypes.SELECT,
-        replacements: {
-          module_ids,
-          event_slug,
-          type
-        },
-      });
-      eventId = eventIdData[0];
-    }
-
-    const module_id = module_ids ?? jobDatas.module_id;
-    const event_id = eventId?.id ?? "";
-    const workflowQuery2 = jobWorkflowQuery(jobDatas.hierarchy_ids);
-    const rows: any[] = await sequelize.query(workflowQuery2, {
-      replacements: { module_id, event_id, program_id, placement_order },
-      type: QueryTypes.SELECT,
-    });
-
-    if (offerData.status?.toLocaleUpperCase() !== 'DRAFT') {
-      if (rows.length > 0 || rows[0]?.levels > 0) {
-        const hasReviewFlow = rows.some(row => row.flow_type.trim() === 'Review');
-        if (offerData.parent_offer_id) {
-          offerData.status = hasReviewFlow ? "Pending Review" : "Pending Approval";
-          const parentOffer = await OfferModel.findByPk(offerData.parent_offer_id);
-          if (parentOffer) {
-            await parentOffer.update({ status: "CLOSED" }, { transaction });
-            await submission?.update({ status: `Counter Offer ${offerData.status}` });
-          }
-        } else {
-          offerData.status = hasReviewFlow ? "Pending Review" : "Pending Approval";
-          await submission?.update({ status: `Offer ${offerData.status}` });
-        }
-      } else {
-        if (offerData.parent_offer_id) {
-          const parentOffer = await OfferModel.findByPk(offerData.parent_offer_id);
-          const oldStatus = parentOffer?.dataValues.status;
-          if (parentOffer) {
-            await parentOffer.update({ status: "CLOSED" }, { transaction });
-            await submission?.update({ status: `Counter Offer Pending Approval` });
-            offerData.status = "Pending Approval";
-            offerData.is_workflow = true;
-            const newStatus = `Counter Offer ${offerData.status}`;
-            const oldData = { status: oldStatus, candidate_id: submission?.dataValues.candidate_id, job_id: submission?.dataValues.job_id, updated_by: userId };
-            const newData = { status: newStatus, candidate_id: submission?.dataValues.candidate_id, job_id: submission?.dataValues.job_id, updated_by: userId };
-            const action = offerData.parent_offer_id ? 'Counter Offer Created' : 'Offer Created';
-            await candidateHistoryService.handleCandidateHistory({ program_id, oldData, newData, action });
-
-          }
-        } else {
-          offerData.status = "Pending Acceptance";
-          await submission?.update({ status: `Offer ${offerData.status}` });
-        }
-      }
-    }
-    const isUpdate = !!offerData.id;
-    const existingOffer = await validateExistingOffer(offerData);
-    if (!isUpdate || (isUpdate && !existingOffer)) {
-      if (existingOffer) {
-        return reply.status(400).send({
-          status_code: 400,
-          trace_id: traceId,
-          message: existingOffer.message,
-          id: existingOffer.id,
-        });
-      }
-    }
-
-    const [newItem] = await OfferModel.upsert(
-      {
-        ...offerData, program_id, vendor_id: submission?.vendor_id, updated_by: userId, created_by: userId, created_on: Date.now(), updated_on: Date.now(),
-        checklist_entity_id: offerData.checklist_entity_id ?? submission?.checklist_entity_id ?? null,
-        checklist_version: offerData.checklist_version ?? submission?.checklist_version ?? null,
-        onboarding_flow_id: submission?.onboarding_flow_id ?? null,
-      },
-      { transaction }
-    );
-    const oldData = {
-      status: null,
-      candidate_id: newItem?.dataValues.candidate_id,
-      job_id: newItem?.dataValues.job_id,
-      updated_by: userId,
-    };
-
-    if (newItem.dataValues.status?.toUpperCase() === "PENDING ACCEPTANCE") {
-      const newData = {
-        candidate_id: newItem?.dataValues.candidate_id,
-        status: newItem?.dataValues.status,
-        job_id: newItem?.dataValues?.job_id,
-        updated_by: newItem?.dataValues.updated_by
-      };
-      await candidateHistoryService.handleCandidateHistory({ program_id, oldData, newData, action: 'Offer Created', });
-      offerNotificationService.processAndSendOfferNotification(
-        token, reply, program_id, newItem, jobDatas, sequelize, user, traceId, offerData, NotificationEventCode.OFFER_RELEASED
-      );
-    } else {
-      offerNotificationService.processAndSendOfferNotification(token, reply, program_id, newItem, jobDatas, sequelize, user, traceId, offerData, NotificationEventCode.OFFER_CREATE);
-      const newData = {
-        candidate_id: newItem?.dataValues.candidate_id,
-        status: newItem?.dataValues.status,
-        job_id: newItem?.dataValues?.job_id,
-        updated_by: newItem?.dataValues.updated_by
-      };
-      await candidateHistoryService.handleCandidateHistory({ program_id, oldData, newData, action: 'Offer Created', });
-    }
-
-    const hierarchyIds = jobs.hierarchy_ids ?? [];
-
-    if (newItem?.onboarding_flow_id) {
-      processOnboardingFlow(jobDatas, newItem, program_id, request, transaction, traceId, reply, hierarchyIds);
-    }
-
-    let jobData = offerData;
-    jobData.userId = userId;
-    jobData.userType = userType;
-
-    let job = { event_title: jobDatas.job_id, job_id: newItem.dataValues.job_id, id: newItem.dataValues.id };
-
-    await sequelize.query(`SELECT job_id FROM jobs WHERE id = :job_id;`, {
-      replacements: { job_id: newItem.dataValues.job_id },
-      type: QueryTypes.SELECT
-    }) as [{ job_id: string }]
-
-    if (offerData.hierarchy && Array.isArray(offerData.hierarchy)) {
-      for (const hierarchyId of offerData.hierarchy) {
-        await OfferHierachy.create(
-          {
-            offer_id: newItem.id,
-            hierarchy: hierarchyId,
-          },
-          { transaction }
-        );
-      }
-    }
-
-    try {
-      await processCustomFields(newItem.id, offerData.custom_fields, transaction);
-      await processFoundationalData(newItem.id, offerData.foundational_data, transaction);
-      const job_max_bill_rate = jobData.financial_details?.rates[0]?.rate_configuration[0]?.base_rate?.clientBillRate?.max_rate;
-      const offer_max_bill_rate = jobData.financial_details?.rates[0]?.rate_configuration[0]?.base_rate?.client_bill_rate;
-      let is_offer_rate_greater = false
-
-      if (Number(offer_max_bill_rate) > Number(job_max_bill_rate)) {
-        is_offer_rate_greater = true;
-      }
-
-      const managers = await fetchManagerIds(offerData);
-
-      jobData.is_offer_rate_greater = is_offer_rate_greater;
-      jobData.job_managers = managers?.job_manager_id;
-      jobData.timesheet_manager_id = managers?.timesheet_manager_id
-      jobData.expense_manager_id = managers?.expense_manager_id
-      jobData.duration = jobData?.financial_details?.billRateValue?.duration_in_days
-      jobData.offer_budget = jobData?.financial_details?.billRateValue?.budget
-      jobDatas.is_offer_rate_greater = is_offer_rate_greater;
-      const weekCount = parseInt(offerData?.financial_details?.billRateValue?.formatted_weeks_days?.match(/(\d+)\s+Weeks?/)?.[1] || "0");
-      jobData.duration = weekCount;
-      jobDatas.duration = weekCount;
-      const workflow = await workflowTriggering(request, reply, program_id, rows, job, jobData, jobDatas, module_name, is_updated, workflow_job_id, event_slug);
-      const offers = await OfferModel.findOne(
-        {
-          where: {
-            id: newItem.id,
-          },
-          transaction
-        });
-
-      if (!workflow) {
-        const workflowQuery2 = ApprovalworkflowQuery(jobDatas.hierarchy_ids);
-        const rows: any[] = await sequelize.query(workflowQuery2, {
-          replacements: { module_id, event_id, program_id, placement_order },
-          type: QueryTypes.SELECT,
-          transaction
-        });
-
-        const workflow = await workflowTriggering(request, reply, program_id, rows, job, jobData, jobDatas, module_name, is_updated, workflow_job_id, event_slug);
-
-        if (!workflow) {
-          if (!offerData.parent_offer_id) {
-            await submission?.update({ status: "Offer Pending Acceptance" }, { transaction });
-            await offers?.update({ status: "Pending Acceptance" }, { transaction });
-          } else {
-            await submission?.update({ status: "Counter Offer Pending Approval" }, { transaction });
-            await offers?.update({ status: "Pending Approval" }, { transaction });
-          }          
-        }
-        else {
-          await submission?.update({ status: "Offer Pending Approval" }, { transaction });
-          await offers?.update({ status: "Pending Approval" }, { transaction });
-        }
-      }
-
-      if (workflow?.workflow?.workflow_status === "completed") {
-        await submission?.update({ status: "Offer Pending Acceptance" }, { transaction });
-        await offers?.update({ status: "Pending Acceptance" }, { transaction });
-      }
-
-    } catch (error) {
-      console.log('error is mwww', error);
-    }
-
-    await transaction.commit();
-
-    reply.status(201).send({
-      status_code: 201,
-      trace_id: traceId,
-      message: "Offer created successfully.",
-      id: newItem?.id,
-    });
-
   } catch (error: any) {
-    if (transaction) {
-      await transaction.rollback();
-    }
-
-    return reply.status(500).send({
-      status_code: 500,
-      trace_id: traceId,
+    return reply.status(error.status_code || 500).send({
+      trace_id: error.trace_id,
       message: error.message,
+      error: error.error || error,
     });
   }
 }
 
 
-async function processOnboardingFlow(job: { job_id: any; id: any; job_template_id: string; }, offer: OfferModel, program_id: string, request: FastifyRequest, transaction: Transaction, traceId: string, reply: FastifyReply, hierarchy_ids: string[]) {
-  const onboarding_flow_id = offer.onboarding_flow_id;
-  const jobTemplateQuery = `
-    SELECT template_name, id, job_id
-    FROM ${config_db}.job_templates
-    WHERE id = :job_template_id
-    LIMIT 1;
-  `;
-  const jobTemplateResult = await sequelize.query(jobTemplateQuery, {
-    replacements: { job_template_id: job.job_template_id },
-    type: QueryTypes.SELECT,
 
-  });
-
-  const jobTemplate: any = jobTemplateResult[0];
-
-  const vendorQuery = `
-          SELECT id, vendor_name, tenant_id
-          FROM ${config_db}.program_vendors
-          WHERE id = :vendor_id
-          `
-
-  const [vendor]: any = await sequelize.query(vendorQuery, {
-    replacements: { vendor_id: offer.vendor_id },
-    type: QueryTypes.SELECT
-  });
-
-  try {
-    let triggers = [OFFER_CREATION, ...getSubsequentTrigger(OFFER_CREATION)];
-
-    const checklistTaskMappings = await getChecklistTaskMappings({ checklist_entity_id: offer.checklist_entity_id, checklist_version: offer.checklist_version, triggers });
-
-    const triggerStepCounts = checklistTaskMappings.reduce((acc: Record<string, number>, ctm: any) => {
-      acc[`${ctm.trigger}_steps_count`]++;
-      return acc;
-    }, getBaseTriggersStepCounts(triggers));
-
-    const workflowUpdates = {
-      associations: {
-        offer_id: offer.id,
-        offer_code: offer.offer_code
-      },
-      attributes: {
-        ...triggerStepCounts
-      }
-    }
-
-    // Call to create workflow steps by checklist task mapping
-    const workflowSteps = await createWorkflowStepsByChecklistTaskMapping(
-      checklistTaskMappings.filter((mapping: any) => mapping.trigger == OFFER_CREATION),
-      program_id,
-      offer.candidate_id,
-      hierarchy_ids,
-      vendor.tenant_id,
-      {
-        job_template_id: jobTemplate.id,
-        job_template_code: jobTemplate.job_id,
-        job_id: job.id,
-        job_code: job.job_id,
-        offer_id: offer.id,
-        offer_code: offer.offer_code,
-      }
-    );
-
-    const onboarding_flow_updates = {
-      workflow: workflowUpdates,
-      steps: workflowSteps
-    }
-
-    try {
-      const credentialing_result = await credentialingService.pushWorkflowUpdatesAndAppendSteps(
-        onboarding_flow_id,
-        onboarding_flow_updates,
-        program_id,
-        request.headers?.authorization!
-      )
-
-      console.log("Workflow updated and possible workflow steps appended successfully in credentialing:", credentialing_result);
-    } catch (error: any) {
-      console.error("Error updating workflow  or appending steps in credentialing service:", error);
-
-      return reply.status(500).send({
-        status_code: 500,
-        trace_id: traceId,
-        message: "Failed to add onboarding steps",
-        error: error.message,
-      });
-    }
-  } catch (error: any) {
-    console.error("Error creating workflow steps:", error);
-
-    return reply.status(500).send({
-      status_code: 500,
-      trace_id: traceId,
-      message: "Failed to create workflow steps.",
-      error: error.message,
-    });
-  }
-}
 
 export async function updateWorkflowReview(
   request: FastifyRequest,
@@ -1004,88 +568,7 @@ export async function fetchUsersBasedOnHierarchy(allPayload: { hierarchy_ids: an
   }
 }
 
-async function validateExistingOffer(offerData: OfferInterface) {
-  let offer;
-  if (offerData.parent_offer_id == null) {
-    const offer = await OfferModel.findOne({
-      where: {
-        job_id: offerData.job_id,
-        candidate_id: offerData.candidate_id,
-        status: {
-          [Op.ne]: "Withdraw",
-        },
-      },
-    });
-    if (offer) {
-      return {
-        message: "Offer already exists for this candidate and job!",
-        id: offer.id,
-      };
-    }
-  } else {
-    offer = await OfferModel.findOne({
-      where: {
-        job_id: offerData.job_id,
-        candidate_id: offerData.candidate_id,
-        parent_offer_id: offerData.parent_offer_id,
-        status: {
-          [Op.notIn]: ["Withdraw", "CLOSED"],
-        }
-      },
-    });
-    if (offer) {
-      return {
-        message: "Counter offer already exists for this candidate and job!",
-        id: offer.id,
-      };
-    }
-  }
-  return null;
-}
 
-async function processCustomFields(
-  offerId: string,
-  customFields: any[],
-  transaction: Transaction
-) {
-  if (Array.isArray(customFields) && customFields.length > 0) {
-    await Promise.all(
-      customFields.map(async (customField) => {
-        await OfferCustomFieldModel.create(
-          {
-            offer_id: offerId,
-            custom_field_id: customField.id,
-            value: customField.value,
-          },
-          { transaction }
-        );
-      })
-    );
-  }
-}
-
-async function processFoundationalData(offerId: string, foundationalData: any[] | undefined, transaction: Transaction) {
-  if (Array.isArray(foundationalData) && foundationalData.length > 0) {
-
-    await OfferMasterDataModel.destroy({
-      where: { offer_id: offerId },
-      transaction,
-    });
-
-    await Promise.all(
-      foundationalData.map(async (data) => {
-        await OfferMasterDataModel.create(
-          {
-            offer_id: offerId,
-            foundation_data_type_id: data.foundation_data_type_id,
-            foundation_data_ids: data.foundation_data_ids,
-          },
-          { transaction }
-        );
-      })
-    );
-  }
-}
 
 export async function updateOffer(request: FastifyRequest, reply: FastifyReply) {
   const traceId = generateCustomUUID();
@@ -1227,37 +710,7 @@ export async function updateOffer(request: FastifyRequest, reply: FastifyReply) 
   }
 }
 
-async function checkDistributionStatusForVendor(program_id: string, userType: string, jobId: any, userData: any[], reply: FastifyReply, traceId: string) {
-  const tenantId = userData?.[0]?.tenant_id;
-  let vendorId: string | undefined;
-  const vendor = await jobRepository.findVendor(program_id, tenantId);
-  vendorId = vendor?.[0]?.id;
-  if (!vendorId) {
-    return reply.status(400).send({
-      status_code: 400,
-      message: "Vendor not found.",
-      trace_id: traceId,
-    });
-  }
 
-  const distribution = await JobDistributionModel.findOne({
-    where: {
-      job_id: jobId,
-      vendor_id: vendorId,
-      program_id,
-    },
-  });
-
-  const distributionStatus = distribution?.status?.toUpperCase() ?? "";
-
-  if (["HOLD", "HALT"].includes(distributionStatus)) {
-    return reply.status(400).send({
-      status_code: 400,
-      message: `Job is currently on '${distributionStatus}', offer cannot be created or updated.`,
-      trace_id: traceId,
-    });
-  }
-}
 
 
 async function approveCounterOffer(offer: any, transaction: any, userId: string) {
